Final version at 31 deg. No watchdog timer working. Can not figure out the bug. The timer works, just not the timer_delay function.

+++ functions.h
@@ -2,7 +2,7 @@
 
 void delay (unsigned long d); 
 
-void timer_delay (long time);
+void timer_delay (unsigned int time);
 
 void initLEDs (void);
